feat: add DCP issuance to MXD
