+++ mxd-runtimes/gradle/catalogserver.versions.toml
@@ -2,7 +2,7 @@
 format.version = "1.1"
 
 [versions]
-edc = "0.10.1"
+edc = "0.12.0-SNAPSHOT"
 tractusx = "0.8.0"
 postgres = "42.7.3"
