+++ mxd-runtimes/tx-catalog-server/build.gradle.kts
@@ -18,19 +18,6 @@ plugins {
     alias(libs.plugins.shadow)
 }
 
-configurations.all {
-    resolutionStrategy {
-        // Tractus-X depends on an earlier version of this, which does not yet contain ConsoleMonitor$Level#defaultLevel()
-        force("org.eclipse.edc:boot-spi:0.10.1")
-        // force the latest version, because the DspHttpCoreExtension now registers the dispatcher WITH the protocol string
-        force("org.eclipse.edc:core-spi:0.10.1")
-        // ContractDefinitionResolver is not present in the older version of EDC used by tractus-Xs
-        force("org.eclipse.edc:catalog-spi:0.10.1")
-        // DSP Remote Message Dispatcher Registry now requires a String argument
-        force("org.eclipse.edc:dsp:0.10.1")
-    }
-}
-
 dependencies {
     // management API - is not included in the FederatedCatalog BOM
     runtimeOnly(catalogLibs.api.management)
@@ -52,7 +39,7 @@ dependencies {
     runtimeOnly(catalogLibs.tx.fc) // file-based node directory
 
     runtimeOnly(libs.edc.did.core) // DidResolverRegistry, DidPublicKeyResolver
-    runtimeOnly(libs.edc.did.web) // for registering the WebDidResolverl
+    runtimeOnly(libs.edc.did.web) // for registering the WebDidResolver
 }
 
 application {
