--- mxd-runtimes/tx-identityhub-sts/src/main/java/org/eclipse/edc/identityhub/demo/IdentityHubExtension.java
@@ -1,31 +1,0 @@
-/*
- *  Copyright (c) 2024 Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
- *
- *  This program and the accompanying materials are made available under the
- *  terms of the Apache License, Version 2.0 which is available at
- *  https://www.apache.org/licenses/LICENSE-2.0
- *
- *  SPDX-License-Identifier: Apache-2.0
- *
- *  Contributors:
- *       Bayerische Motoren Werke Aktiengesellschaft (BMW AG) - initial API and implementation
- *
- */
-
-package org.eclipse.edc.identityhub.demo;
-
-import org.eclipse.edc.identityhub.spi.transformation.ScopeToCriterionTransformer;
-import org.eclipse.edc.runtime.metamodel.annotation.Extension;
-import org.eclipse.edc.runtime.metamodel.annotation.Provider;
-import org.eclipse.edc.spi.system.ServiceExtension;
-
-
-@Extension("DCP Demo: Core Extension for IdentityHub")
-public class IdentityHubExtension implements ServiceExtension {
-
-    @Provider
-    public ScopeToCriterionTransformer createScopeTransformer() {
-        return new TxScopeToCriterionTransformer();
-    }
-
-}
