--- mxd-runtimes/tx-identityhub-sts/src/main/resources/META-INF/services/org.eclipse.edc.spi.system.ServiceExtension
@@ -1,16 +1,0 @@
-#
-#  Copyright (c) 2024 Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
-#
-#  This program and the accompanying materials are made available under the
-#  terms of the Apache License, Version 2.0 which is available at
-#  https://www.apache.org/licenses/LICENSE-2.0
-#
-#  SPDX-License-Identifier: Apache-2.0
-#
-#  Contributors:
-#       Bayerische Motoren Werke Aktiengesellschaft (BMW AG) - initial API and implementation
-#
-#
-
-org.eclipse.edc.identityhub.demo.IdentityHubExtension
-org.eclipse.edc.identityhub.seed.SuperUserSeedExtension
