+++ mxd-runtimes/tx-issuerservice/src/main/java/org/eclipse/edc/issuerservice/demo/attestation/DemoAttestationsExtension.java
@@ -1,0 +1,46 @@
+/*
+ *  Copyright (c) 2025 Cofinity-X
+ *
+ *  This program and the accompanying materials are made available under the
+ *  terms of the Apache License, Version 2.0 which is available at
+ *  https://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  SPDX-License-Identifier: Apache-2.0
+ *
+ *  Contributors:
+ *       Cofinity-X - initial API and implementation
+ *
+ */
+
+package org.eclipse.edc.issuerservice.demo.attestation;
+
+import org.eclipse.edc.issuerservice.spi.issuance.attestation.AttestationDefinitionValidatorRegistry;
+import org.eclipse.edc.issuerservice.spi.issuance.attestation.AttestationSourceFactoryRegistry;
+import org.eclipse.edc.runtime.metamodel.annotation.Extension;
+import org.eclipse.edc.runtime.metamodel.annotation.Inject;
+import org.eclipse.edc.spi.system.ServiceExtension;
+import org.eclipse.edc.spi.system.ServiceExtensionContext;
+
+
+@Extension(value = DemoAttestationsExtension.NAME)
+public class DemoAttestationsExtension implements ServiceExtension {
+
+    public static final String NAME = "Demo Attestations Extension";
+
+    @Inject
+    private AttestationSourceFactoryRegistry registry;
+
+    @Inject
+    private AttestationDefinitionValidatorRegistry validatorRegistry;
+
+    @Override
+    public String name() {
+        return NAME;
+    }
+
+    @Override
+    public void initialize(ServiceExtensionContext context) {
+        registry.registerFactory("demo", new DemoAttestationSourceFactory());
+        validatorRegistry.registerValidator("demo", new DemoAttestatonSourceValidator());
+    }
+}
