+++ mxd/bob.tf
@@ -26,9 +26,9 @@ module "bob-connector" {
   }
   dcp-config = {
     id                     = var.bob-did
-    sts_token_url          = "http://${var.bob-identityhub-host}:7084/api/credentials/token"
+    sts_token_url          = "http://bob-ih:7084/api/sts/token"
     sts_client_id          = var.bob-did
-    sts_clientsecret_alias = "participant-bob-sts-client-secret"
+    sts_clientsecret_alias = "${var.bob-did}-sts-client-secret"
   }
   dataplane = {
     privatekey-alias = "${var.bob-did}#signing-key-1"
@@ -45,10 +45,11 @@ module "bob-connector" {
     password = module.bob-minio.minio-password
     url      = module.bob-minio.minio-url
   }
+  useSVE = var.useSVE
 }
 
 module "bob-identityhub" {
-  depends_on = [module.bob-connector]
+  depends_on = [module.bob-connector] // depends because of the vault
   source     = "./modules/identity-hub"
   database = {
     user     = local.databases.bob.database-username
