+++ mxd/modules/identity-hub/outputs.tf
@@ -29,3 +29,7 @@ output "ports" {
 output "ih-superuser-apikey" {
   value = var.ih_superuser_apikey
 }
+
+output "sts-token-url" {
+  value = "http://${kubernetes_service.ih-service.metadata.0.name}:${var.ports.sts-api}${var.sts-token-path}"
+}
