+++ mxd/modules/identity-hub/services.tf
@@ -23,12 +23,12 @@ resource "kubernetes_service" "ih-service" {
     }
     # we need a stable IP, otherwise there will be a cycle with the issuer
     port {
-      name = "presentation"
-      port = var.ports.presentation-api
+      name = "credentials"
+      port = var.ports.credentials-api
     }
     port {
       name = "debug"
-      port = var.ports.ih-debug
+      port = var.ports.debug
     }
     port {
       name = "identity"
@@ -40,7 +40,7 @@ resource "kubernetes_service" "ih-service" {
     }
     port {
       name = "sts"
-      port = var.ports.ih-sts
+      port = var.ports.sts-api
     }
   }
 }
