+++ mxd/modules/issuerservice/ingress.tf
@@ -1,20 +1,14 @@
 #
-#  Copyright (c) 2023 Contributors to the Eclipse Foundation
-#
-#  See the NOTICE file(s) distributed with this work for additional
-#  information regarding copyright ownership.
+#  Copyright (c) 2025 Cofinity-X
 #
 #  This program and the accompanying materials are made available under the
 #  terms of the Apache License, Version 2.0 which is available at
 #  https://www.apache.org/licenses/LICENSE-2.0
 #
-#  Unless required by applicable law or agreed to in writing, software
-#  distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
-#  WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
-#  License for the specific language governing permissions and limitations
-#  under the License.
-#
 #  SPDX-License-Identifier: Apache-2.0
+#
+#  Contributors:
+#       Cofinity-X - initial API and implementation
 #
 
 resource "kubernetes_ingress_v1" "api-ingress" {
@@ -30,26 +24,26 @@ resource "kubernetes_ingress_v1" "api-ingress" {
     ingress_class_name = "nginx"
     rule {
       http {
-        # ingress path for the IdentityHub Identity (=management) API
+
         path {
-          path = "/${var.url-path}/cs(/|$)(.*)"
+          path = "/issuer/cs(/|$)(.*)"
           backend {
             service {
-              name = kubernetes_service.ih-service.metadata.0.name
+              name = kubernetes_service.issuerservice-service.metadata.0.name
               port {
-                number = var.ports.ih-identity-api
+                number = var.ports.identity
               }
             }
           }
         }
-        # ingress path for the STS Token API
+
         path {
-          path = "/${var.url-path}/sts(/|$)(.*)"
+          path = "/issuer/ad(/|$)(.*)"
           backend {
             service {
-              name = kubernetes_service.ih-service.metadata.0.name
+              name = kubernetes_service.issuerservice-service.metadata.0.name
               port {
-                number = var.ports.sts-api
+                number = var.ports.issueradmin
               }
             }
           }
@@ -62,10 +56,10 @@ resource "kubernetes_ingress_v1" "api-ingress" {
 // the DID endpoint can not actually modify the URL, otherwise it'll mess up the DID resolution
 resource "kubernetes_ingress_v1" "did-ingress" {
   metadata {
-    name      = "${var.url-path}-did-server-ingress"
+    name      = "${var.humanReadableName}-did-ingress"
     namespace = var.namespace
     annotations = {
-      "nginx.ingress.kubernetes.io/rewrite-target" = "/${var.url-path}/$2"
+      "nginx.ingress.kubernetes.io/rewrite-target" = "/issuer/$2"
     }
   }
 
@@ -75,12 +69,12 @@ resource "kubernetes_ingress_v1" "did-ingress" {
       http {
         # ingress routes for the DID endpoint
         path {
-          path = "/${var.url-path}(/|&)(.*)"
+          path = "/issuer(/|&)(.*)"
           backend {
             service {
-              name = kubernetes_service.ih-service.metadata.0.name
+              name = kubernetes_service.issuerservice-service.metadata.0.name
               port {
-                number = var.ports.ih-did
+                number = var.ports.did
               }
             }
           }
