+++ mxd/modules/issuerservice/outputs.tf
@@ -1,0 +1,19 @@
+#
+#  Copyright (c) 2025 Cofinity-X
+#
+#  This program and the accompanying materials are made available under the
+#  terms of the Apache License, Version 2.0 which is available at
+#  https://www.apache.org/licenses/LICENSE-2.0
+#
+#  SPDX-License-Identifier: Apache-2.0
+#
+#  Contributors:
+#       Cofinity-X - initial API and implementation
+
+
+output "endpoints" {
+  value = {
+    admin    = "${kubernetes_service.issuerservice-service.metadata.0.name}:${var.ports.issueradmin}"
+    identity = "${kubernetes_service.issuerservice-service.metadata.0.name}:${var.ports.identity}"
+  }
+}
