--- mxd/modules/sts/services.tf
@@ -1,38 +1,0 @@
-#
-#  Copyright (c) 2024 Metaform Systems, Inc.
-#
-#  This program and the accompanying materials are made available under the
-#  terms of the Apache License, Version 2.0 which is available at
-#  https://www.apache.org/licenses/LICENSE-2.0
-#
-#  SPDX-License-Identifier: Apache-2.0
-#
-#  Contributors:
-#       Metaform Systems, Inc. - initial API and implementation
-#
-
-resource "kubernetes_service" "sts-service" {
-  metadata {
-    name      = var.humanReadableName
-    namespace = var.namespace
-  }
-  spec {
-    type = "NodePort"
-    selector = {
-      App = kubernetes_deployment.sts.spec.0.template.0.metadata[0].labels.App
-    }
-    # we need a stable IP, otherwise there will be a cycle with the issuer
-    port {
-      name = "accounts"
-      port = var.ports.accounts
-    }
-    port {
-      name = "sts"
-      port = var.ports.sts
-    }
-    port {
-      name = "debug"
-      port = var.ports.debug
-    }
-  }
-}
