+++ mxd/modules/vault/vault-values.yaml
@@ -1,0 +1,22 @@
+#
+#  Copyright (c) 2024 Metaform Systems, Inc.
+#
+#  This program and the accompanying materials are made available under the
+#  terms of the Apache License, Version 2.0 which is available at
+#  https://www.apache.org/licenses/LICENSE-2.0
+#
+#  SPDX-License-Identifier: Apache-2.0
+#
+#  Contributors:
+#       Metaform Systems, Inc. - initial API and implementation
+#
+
+server:
+  postStart:
+hashicorp:
+  timeout: 30
+  healthCheck:
+    enabled: true
+    standbyOk: true
+  paths:
+    secret: /v1/secret
